[APM] Correct timezone formatting
* Don't set the timezone anywhere in APM since it's already set in autoload
* Always use the local timezone for the chart tooltip

This makes the timezone handling consitently use the users set preference _except_ in the charts where the local time is always used.

In the charts we have to go through react-vis and d3-scale and nothing I tried to modify the scale on the x-axis ever really worked. We could revisit this later.

Fixes #47832
Fixes #48355